Split 'user.go' into files
