+++ internal/server/api/v1/users/update_test.go
@@ -54,7 +54,7 @@ func TestUser_UpdateUser(t *testing.T) {
 
 	gin.SetMode(gin.ReleaseMode)
 	r := gin.New()
-	r.PATCH("/users/:id", NewUserAPI(m).UpdateUser)
+	r.PATCH("/users/:id", NewUserAPI(m).Update)
 
 	p, _ := json.Marshal(input.Payload)
 	req, _ := http.NewRequest("PATCH", fmt.Sprintf("/users/%d", input.ID), bytes.NewBuffer(p))
